+++ spec/models/maths_and_physics/school_eligibility_spec.rb
@@ -14,6 +14,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it has an education phase of secondary or middle deemed secondary" do
           let(:school_attributes) { college_attributes.merge({phase: :middle_deemed_secondary}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { college_attributes.merge({phase: :middle_deemed_secondary, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
 
         context "and it has an education phase of nursery, primary, middle deemed primary, 16+ or all through" do
@@ -28,6 +33,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it has an education phase of secondary or middle deemed secondary" do
           let(:school_attributes) { academy_school_attributes.merge({phase: :secondary}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { academy_school_attributes.merge({phase: :secondary, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
 
         context "and it has an education phase of nursery, primary, middle deemed primary, 16+ or all through" do
@@ -42,6 +52,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it has an education phase of secondary or middle deemed secondary" do
           let(:school_attributes) { free_school_attributes.merge({phase: :secondary}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { free_school_attributes.merge({phase: :secondary, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
 
         context "and it has an education phase of nursery, primary, middle deemed primary, 16+ or all through" do
@@ -56,6 +71,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it has an education phase of secondary or middle deemed secondary" do
           let(:school_attributes) { la_maintained_attributes.merge({phase: :secondary}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { la_maintained_attributes.merge({phase: :secondary, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
 
         context "and it has an education phase of nursery, primary, middle deemed primary, 16+ or all through" do
@@ -70,6 +90,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it has an education phase of secondary or middle deemed secondary" do
           let(:school_attributes) { special_school_attributes.merge({phase: :secondary}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { special_school_attributes.merge({phase: :secondary, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
 
         context "and it has an education phase of nursery, primary, middle deemed primary, 16+ or all through" do
@@ -80,6 +105,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it doesn't have an education phase" do
           let(:school_attributes) { special_school_attributes.merge({phase: :not_applicable}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { special_school_attributes.merge({phase: :not_applicable, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
       end
 
@@ -89,6 +119,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it has an education phase of secondary or middle deemed secondary" do
           let(:school_attributes) { special_free_school_attributes.merge({phase: :secondary}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { special_free_school_attributes.merge({phase: :secondary, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
 
         context "and it has an education phase of nursery, primary, middle deemed primary, 16+ or all through" do
@@ -99,6 +134,11 @@ RSpec.describe MathsAndPhysics::SchoolEligibility do
         context "and it doesn't have an education phase" do
           let(:school_attributes) { special_free_school_attributes.merge({phase: :not_applicable}) }
           it { is_expected.to be true }
+
+          context "and it is closed" do
+            let(:school_attributes) { special_free_school_attributes.merge({phase: :not_applicable, close_date: Date.yesterday}) }
+            it { is_expected.to be false }
+          end
         end
       end